+++ GameEngine/Snowflake.cs
@@ -4,12 +4,12 @@ internal class Snowflake
     /// <summary>
     /// Свойство положения снежинки по x
     /// </summary>
-    public  int X { get; set; }
+    public  float X { get; set; }
 
     /// <summary>
     /// Свойство положения снежинки по y
     /// </summary>
-    public int Y { get; set; }
+    public float Y { get; set; }
 
     /// <summary>
     /// Свойство длины стороны снежинки
